+++ cutplace/ranges.py
@@ -407,7 +407,7 @@ class Range(object):
                 item_index += 1
             if not is_valid:
                 raise errors.RangeValueError(
-                    "%s is %r but must be within range: %r" % (name, value, self), location)
+                    "%s is %r but must be within range: %s" % (name, value, self), location)
 
 
 class DecimalRange(Range):