+++ docs/changes.rst
@@ -5,6 +5,21 @@ Revision history
 This chapter describes improvements compared to earlier versions of cutplace.
 
 
+Version 0.8.4, 2015-02-xx
+=========================
+
+* Fixed reading of non ASCII values from ODS under Python 2.
+* Improved documentation: cleaned up section on :ref:`exit-code`.
+* Improved API documentation: added a section on :ref:`writing-data`.
+* Improved API: changed validation of length for fixed field values:
+  :py:class:`cutplace.Writer` rejects too long values with a
+  :py:exc:`~cutplace.errors.FieldValueError` and automatically pads too short
+  values with trailing blanks while the low level
+  :py:class:`cutplace.rowio.FixedRowWriter` rejects both cases with an
+  :py:exc:`AssertionError`. Furthermore the length of fixed values is now
+  checked before validating it against the field format rule.
+
+
 Version 0.8.3, 2015-01-31
 =========================